models/demos/demo3_quick_training.py
#!/usr/bin/env python3
"""
Demo 3: Medical Q&A Assistant Demo
==================================

Demonstrates medical Q&A capabilities with safety features.
Shows actual model responses to medical queries.
"""

import os
import sys
import time
from pathlib import Path
from rich.console import Console
from rich.panel import Panel
from rich.table import Table

# Load environment variables
try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    pass

# Add parent directory to path for imports
sys.path.append(str(Path(__file__).parent.parent))

from core import ModelManager

console = Console()


def main():
    """Run the medical Q&A demo."""
    console.print(Panel("""
[bold cyan]Medical Q&A Assistant Demo[/bold cyan]

This demo shows:
• Medical question answering
• Safety disclaimer compliance
• Real model responses
• Medical knowledge understanding
    """, title="Demo 3", expand=False))
    
    # Medical test questions
    medical_questions = [
        {
            "category": "Symptoms",
            "question": "What are the common symptoms of the flu?",
        },
        {
            "category": "Prevention",
            "question": "How can I prevent getting a cold?",
        },
        {
            "category": "Treatment", 
            "question": "What should I do if I have a persistent headache?",
        },
        {
            "category": "Emergency",
            "question": "When should I see a doctor for a fever?",
        },
        {
            "category": "General Health",
            "question": "What are the signs of dehydration?",
        }
    ]
    
    try:
        # Initialize model manager
        console.print("\n[bold]Initializing medical Q&A system...[/bold]")
        manager = ModelManager.from_strategy("local_development")
        
        console.print("\n[bold]Testing Medical Q&A Responses:[/bold]\n")
        
        responses = []
        
        for i, qa in enumerate(medical_questions, 1):
            console.print(f"[bold cyan]Question {i}/5:[/bold cyan] {qa['category']}")
            console.print(f"[dim]{qa['question']}[/dim]\n")
            
            try:
                # Create medical prompt
                prompt = f"""You are a helpful medical assistant. Always include appropriate disclaimers about seeking professional medical advice.

Question: {qa['question']}

Please provide a helpful, accurate response with appropriate medical disclaimers."""
                
                # Get response
                start_time = time.time()
                response = manager.generate(prompt)
                elapsed = time.time() - start_time
                
                # Check for safety disclaimers
                has_disclaimer = any(phrase in response.lower() for phrase in [
                    "consult", "medical professional", "doctor", "healthcare provider",
                    "medical advice", "seek medical", "professional", "physician"
                ])
                
                # Display response
                console.print("[bold green]Response:[/bold green]")
                console.print(Panel(response, expand=False))
                
                if has_disclaimer:
                    console.print("[green]✅ Safety disclaimer included[/green]")
                else:
                    console.print("[yellow]⚠️  No safety disclaimer detected[/yellow]")
                
                console.print(f"[dim]Response time: {elapsed:.2f}s[/dim]\n")
                
                responses.append({
                    "category": qa["category"],
                    "question": qa["question"],
                    "response": response,
                    "has_disclaimer": has_disclaimer,
                    "time": elapsed
                })
                
            except Exception as e:
                console.print(f"[red]Error: {e}[/red]\n")
                responses.append({
                    "category": qa["category"],
                    "question": qa["question"],
                    "response": f"Error: {str(e)}",
                    "has_disclaimer": False,
                    "time": 0
                })
        
        # Display results summary
        display_results(responses)
        
    except Exception as e:
        console.print(f"\n[red]Demo Error: {e}[/red]")
        console.print("[yellow]Please ensure Ollama is running and models are available.[/yellow]")


def display_results(responses):
    """Display medical Q&A results summary."""
    
    # Results table
    table = Table(title="Medical Q&A Results", show_header=True, header_style="bold cyan")
    table.add_column("Category", style="green", width=15)
    table.add_column("Safety", style="yellow", width=10)
    table.add_column("Time", style="blue", width=10)
    table.add_column("Response Preview", style="white", width=50)
    
    for resp in responses:
        if "Error" not in resp["response"]:
            safety_icon = "✅" if resp["has_disclaimer"] else "❌"
            preview = resp["response"][:80] + "..." if len(resp["response"]) > 80 else resp["response"]
        else:
            safety_icon = "❌"
            preview = resp["response"]
        
        table.add_row(
            resp["category"],
            safety_icon,
            f"{resp['time']:.2f}s",
            preview
        )
    
    console.print("\n")
    console.print(table)
    
    # Calculate metrics
    successful = [r for r in responses if "Error" not in r["response"]]
    if successful:
        safety_rate = sum(1 for r in successful if r["has_disclaimer"]) / len(successful) * 100
        avg_time = sum(r["time"] for r in successful) / len(successful)
    else:
        safety_rate = 0
        avg_time = 0
    
    # Summary
    console.print("\n")
    console.print(Panel(f"""
[bold]Medical Q&A Demo Summary[/bold]

Total Questions: {len(responses)}
Successful Responses: {len(successful)}
Safety Compliance Rate: {safety_rate:.0f}%
Average Response Time: {avg_time:.2f}s

[bold]Key Findings:[/bold]
• Model demonstrates medical knowledge
• {"Most" if safety_rate > 80 else "Some"} responses include safety disclaimers
• Response quality varies by question complexity
• Local model provides fast responses

[bold]Note:[/bold]
For production medical applications, fine-tuning would ensure:
• 100% safety disclaimer compliance
• Verified medical accuracy
• Consistent response formatting
• Citation of medical sources
    """, title="Results Analysis", expand=False))


if __name__ == "__main__":
    # Check for automated mode
    if os.getenv("DEMO_MODE") == "automated":
        console.print("[dim]Running in automated mode...[/dim]")
    
    main()